+++ include/onnxruntime/core/framework/data_types_internal.h
@@ -13,9 +13,11 @@
 #include "boost/mp11.hpp"
 
 #include "core/common/common.h"
+#ifndef SHARED_PROVIDER
 #include "core/common/type_list.h"
 #include "core/framework/data_types.h"
 #include "core/graph/onnx_protobuf.h"
+#endif
 
 namespace onnxruntime {
 namespace utils {
